Problem with New Configure Script (posted 4/8/11)
Problem: There is a problem with the new configure script on Opensuse and
Ubuntu Linux. It will fail on creating configure.wrf file, and may result
in file removal in the top WRFV3/ directory. If configure.wrf is
successfully created, then this fix doesn't affect you.

Solution: If files are removed from WRFV3 directory, you will have to
redownload the V3.3 tar file, or re-untar the file. Then download the
configure_fix.tar file in the WRFV3 directory. This file contains a new
configure script, and a new arch/configure_new.default file.